Uploding lucid Vx42.
This is a project which was worked on earlier without as vcs and hence project is unorganized and requires lot of cleanup.
Current state of project include the project files when the lucid was at Vx42.
No documentation is available on how to work on this project.

Need to try orginize the project and do further development.